+++ src/storage/mod.rs
@@ -5,7 +5,9 @@ mod snapshot;
 mod stored_value;
 mod value;
 
-pub(crate) use in_memory::{ExpirationUpdate, InMemoryStore, SetCondition, SetExpiration};
+pub(crate) use in_memory::{
+    ExpirationUpdate, ExpireCondition, InMemoryStore, SetCondition, SetExpiration,
+};
 pub(crate) use snapshot::{SnapshotDataError, SnapshotEntry, SnapshotValue};
 
 #[cfg(test)]